changes (BREAKING) rename package name from :feature:movies in order to follow a feature module name standard.
