+++ app/src/main/java/com/christopher_elias/movies/MoviesApp.kt
@@ -1,7 +1,8 @@
 package com.christopher_elias.movies
 
 import android.app.Application
-import com.christopher_elias.actors.di.featureActorsModule
+import com.christopher_elias.features.actors.di.featureActorsModule
+import com.christopher_elias.features.movies.di.featureMoviesModule
 import com.christopher_elias.movies.di.*
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
