--- features/actors/src/main/java/com/christopher_elias/actors/presentation/detail/ActorDetailViewModel.kt
@@ -1,12 +1,0 @@
-package com.christopher_elias.actors.presentation.detail
-
-/*
- * Created by Christopher Elias on 2/05/2021
- * christopher.mike.96@gmail.com
- *
- * Loop Ideas
- * Lima, Peru.
- */
-
-class ActorDetailViewModel {
-}
