+++ features/actors/src/main/java/com/christopher_elias/features/actors/domain/repository/ActorsRepository.kt
@@ -1,0 +1,19 @@
+package com.christopher_elias.features.actors.domain.repository
+
+import com.christopher_elias.features.actors.domain.models.Actor
+import com.christopher_elias.functional_programming.Either
+import com.christopher_elias.functional_programming.Failure
+
+/*
+ * Created by Christopher Elias on 2/05/2021
+ * christopher.mike.96@gmail.com
+ *
+ * Loop Ideas
+ * Lima, Peru.
+ */
+
+interface ActorsRepository {
+
+    suspend fun getActors(): Either<Failure, List<Actor>>
+
+}
