+++ features/actors/src/main/java/com/christopher_elias/features/actors/presentation/detail/ActorDetailViewModel.kt
@@ -1,0 +1,12 @@
+package com.christopher_elias.features.actors.presentation.detail
+
+/*
+ * Created by Christopher Elias on 2/05/2021
+ * christopher.mike.96@gmail.com
+ *
+ * Loop Ideas
+ * Lima, Peru.
+ */
+
+class ActorDetailViewModel {
+}
